Make `ColumnMarker` a static nested class.
